refactored TEMPLATE static variable to be less ambiguous
