+++ testing/testing.client.side/src/test/java/com/peregrine/it/admin/DeleteNodeIT.java
@@ -33,7 +33,7 @@ import static com.peregrine.commons.util.PerConstants.PAGE_CONTENT_TYPE;
 import static com.peregrine.commons.util.PerConstants.PAGE_PRIMARY_TYPE;
 import static com.peregrine.commons.util.PerConstants.SLING_ORDERED_FOLDER;
 import static com.peregrine.commons.util.PerConstants.SLING_RESOURCE_TYPE;
-import static com.peregrine.commons.util.PerUtil.TEMPLATE;
+import static com.peregrine.commons.util.PerConstants.TEMPLATE;
 import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
